Changed RAS Infrastructure.
All RAS classes will split into Listener class, which is a subclass of
GK_RASListener and a worker Class derived from GK_RASWorker. These are
defined in RasListener.h and RasWorker.h, see there for details.

The RasWorker derives from PThread and will create a new thread for each
PDU. This thread will auto terminate when the processing of the PDU is
finished. With this implementation the listening socket in the RasListener
will not be blocked for a too long time.