Fix compilation with system libpt and libh323plus
When libpt and libh323plus are installed as system packages, PT_LIBDIR and OH323_LIBDIR do not reflect the actual location of these libraries, which makes version.cxx compilation fail due to missing dependencies. Drop these formal dependencies.

gnugk linking process is unaffected; it will succeed as long as these libraries are in default locations for the system.